requirements.txt
BareNecessities==0.2.8
beautifulsoup4==4.6.0
certifi==2018.4.16
chardet==3.0.4
CommandTool==0.5.3
dj-database-url==0.5.0
Django==2.0.7
django-activeurl==0.1.12
django-appconf==1.0.2
django-bootstrap4==0.0.6
django-ckeditor==5.5.0
django-classy-tags==0.8.0
django-cleanup==2.1.0
django-cms==3.5.2
django-crispy-forms==1.7.2
django-formtools==2.1
django-imagekit==4.0.2
django-inplaceedit==1.4.1
django-js-asset==1.1.0
django-material==1.2.5
django-menu==0.1.12
django-modeltranslation==0.12.2
django-sekizai==0.10.0
django-translation-manager==1.0.0
django-treebeard==4.3
djangocms-admin-style==1.2.8
djangocms-text-ckeditor==3.6.0
djangorestframework==3.8.2
Gallery==0.1.0
html5lib==0.9999999
idna==2.7
lxml==4.2.3
markdownify==0.4.1
pilkit==2.0
Pillow==5.2.0
polib==1.1.0
pprint==0.1
psycopg2==2.7.5
python-gettext==3.0
pytz==2018.4
requests==2.19.1
six==1.11.0
South==1.0.2
urllib3==1.23